requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alepython @ git+https://github.com/blent-ai/alepython.git@286350ab674980a32270db2a0b5ccca1380312a7
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.135.1
fastjsonschema==2.21.2
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
imbalanced-learn==0.14.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==3.13.2
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
namex==0.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.3
notebook_shim==0.2.4
numba==0.64.0
numpy==2.4.2
opt_einsum==3.4.0
optree==0.19.0
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==2.1.0
sentence-transformers==5.2.3
shap==0.50.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
smmap==5.0.2
soundfile==0.13.1
soupsieve==2.8.3
stack-data==0.6.3
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.2.20
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
wfdb==4.3.1
win32_setctime==1.2.0
wrapt==2.1.1
yarl==1.22.0
